+++ xvidcore/src/image/qpel.h
@@ -19,7 +19,7 @@
  *  along with this program ; if not, write to the Free Software
  *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307 USA
  *
- * $Id: qpel.h,v 1.2 2004-03-22 22:36:23 edgomez Exp $
+ * $Id: qpel.h,v 1.2.2.1 2004-10-12 21:06:34 edgomez Exp $
  *
  ****************************************************************************/
 
@@ -116,8 +116,8 @@ new_interpolate16x16_quarterpel(uint8_t * const cur,
 
 	int32_t x_int, y_int;
 
-	const int32_t xRef = x*4 + dx;
-	const int32_t yRef = y*4 + dy;
+	const int32_t xRef = (int)x*4 + dx;
+	const int32_t yRef = (int)y*4 + dy;
 
 	Ops = xvid_QP_Funcs; /* TODO: pass as argument */
 	quads = (dx&3) | ((dy&3)<<2);
@@ -131,7 +131,7 @@ new_interpolate16x16_quarterpel(uint8_t * const cur,
 		y_int--;
 
 	dst = cur + y * stride + x;
-	src = refn + y_int * stride + x_int;
+	src = refn + y_int * (int)stride + x_int;
 
 	tmp = refh; /* we need at least a 16 x stride scratch block */
 
@@ -218,8 +218,8 @@ new_interpolate16x8_quarterpel(uint8_t * const cur,
 
 	int32_t x_int, y_int;
 
-	const int32_t xRef = x*4 + dx;
-	const int32_t yRef = y*4 + dy;
+	const int32_t xRef = (int)x*4 + dx;
+	const int32_t yRef = (int)y*4 + dy;
 
 	Ops = xvid_QP_Funcs; /* TODO: pass as argument */
 	quads = (dx&3) | ((dy&3)<<2);
@@ -233,7 +233,7 @@ new_interpolate16x8_quarterpel(uint8_t * const cur,
 		y_int--;
 
 	dst = cur + y * stride + x;
-	src = refn + y_int * stride + x_int;
+	src = refn + y_int * (int)stride + x_int;
 
 	tmp = refh; /* we need at least a 16 x stride scratch block */
 
@@ -318,8 +318,8 @@ new_interpolate8x8_quarterpel(uint8_t * const cur,
 
 	int32_t x_int, y_int;
 
-	const int32_t xRef = x*4 + dx;
-	const int32_t yRef = y*4 + dy;
+	const int32_t xRef = (int)x*4 + dx;
+	const int32_t yRef = (int)y*4 + dy;
 
 	Ops = xvid_QP_Funcs; /* TODO: pass as argument */
 	quads = (dx&3) | ((dy&3)<<2);
@@ -333,7 +333,7 @@ new_interpolate8x8_quarterpel(uint8_t * const cur,
 		y_int--;
 
 	dst = cur + y * stride + x;
-	src = refn + y_int * stride + x_int;
+	src = refn + y_int * (int)stride + x_int;
 
 	tmp = refh; /* we need at least a 16 x stride scratch block */
